읭? 다시 깃 init 했는데 이러면 큰일나는겨? 그렇게 깃으로 수많은 코드를 날려먹고도 정신못차리는분?

+++ all_made_structed.c
@@ -181,7 +181,7 @@ struct skill
     int use_mp;
 };
 
-struct skill set_skill(struct skill s, char s_name[],int get_level ,int type,int magic_rate,int use_mp)
+struct skill set_skill(struct skill s, char s_name[],int get_level ,int type,float magic_rate,int use_mp)
 {
     strcpy(s.name, s_name);
     s.get_level = get_level;
